backend/requirements.txt
boto3>=1.28.0
python-dotenv>=0.19.0
tabulate>=0.8.9
python-dotenv